// src/api/__generated__/story/drama/drama.ts
export type playResponse200 = {
  data: Play200;
  status: 200;
};

export type playResponse400 = {
  data: ResultObject;
  status: 400;
};

export type playResponseSuccess = playResponse200 & {
  headers: Headers;
};
export type playResponseError = playResponse400 & {
  headers: Headers;
};

export type playResponse = playResponseSuccess | playResponseError;
// ...
export type toggleLikeResponse200 = {
  data: ToggleLike200;
  status: 200;
};

export type toggleLikeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type toggleLikeResponseSuccess = toggleLikeResponse200 & {
  headers: Headers;
};
export type toggleLikeResponseError = toggleLikeResponse400 & {
  headers: Headers;
};

export type toggleLikeResponse =
  | toggleLikeResponseSuccess
  | toggleLikeResponseError;
// ...
export type toggleFavoriteResponse200 = {
  data: ToggleFavorite200;
  status: 200;
};

export type toggleFavoriteResponse400 = {
  data: ResultObject;
  status: 400;
};

export type toggleFavoriteResponseSuccess = toggleFavoriteResponse200 & {
  headers: Headers;
};
export type toggleFavoriteResponseError = toggleFavoriteResponse400 & {
  headers: Headers;
};

export type toggleFavoriteResponse =
  | toggleFavoriteResponseSuccess
  | toggleFavoriteResponseError;
// ...
export type completeResponse200 = {
  data: Complete200;
  status: 200;
};

export type completeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type completeResponseSuccess = completeResponse200 & {
  headers: Headers;
};
export type completeResponseError = completeResponse400 & {
  headers: Headers;
};

export type completeResponse = completeResponseSuccess | completeResponseError;
// ...
export type postReviewResponse200 = {
  data: ReviewResponse;
  status: 200;
};

export type postReviewResponse400 = {
  data: ResultObject;
  status: 400;
};

export type postReviewResponseSuccess = postReviewResponse200 & {
  headers: Headers;
};
export type postReviewResponseError = postReviewResponse400 & {
  headers: Headers;
};

export type postReviewResponse =
  | postReviewResponseSuccess
  | postReviewResponseError;
// ...
export type playEpisodeResponse200 = {
  data: PlayEpisode200;
  status: 200;
};

export type playEpisodeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type playEpisodeResponseSuccess = playEpisodeResponse200 & {
  headers: Headers;
};
export type playEpisodeResponseError = playEpisodeResponse400 & {
  headers: Headers;
};

export type playEpisodeResponse =
  | playEpisodeResponseSuccess
  | playEpisodeResponseError;
// ...
export type toggleFavoriteDramaResponse200 = {
  data: ToggleFavoriteDrama200;
  status: 200;
};

export type toggleFavoriteDramaResponse400 = {
  data: ResultObject;
  status: 400;
};

export type toggleFavoriteDramaResponseSuccess =
  toggleFavoriteDramaResponse200 & {
    headers: Headers;
  };
export type toggleFavoriteDramaResponseError =
  toggleFavoriteDramaResponse400 & {
    headers: Headers;
  };

export type toggleFavoriteDramaResponse =
  | toggleFavoriteDramaResponseSuccess
  | toggleFavoriteDramaResponseError;
// ...
export type toggleLikeEpisodeResponse200 = {
  data: ToggleLikeEpisode200;
  status: 200;
};

export type toggleLikeEpisodeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type toggleLikeEpisodeResponseSuccess = toggleLikeEpisodeResponse200 & {
  headers: Headers;
};
export type toggleLikeEpisodeResponseError = toggleLikeEpisodeResponse400 & {
  headers: Headers;
};

export type toggleLikeEpisodeResponse =
  | toggleLikeEpisodeResponseSuccess
  | toggleLikeEpisodeResponseError;
// ...
export type completeEpisodeResponse200 = {
  data: CompleteEpisode200;
  status: 200;
};

export type completeEpisodeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type completeEpisodeResponseSuccess = completeEpisodeResponse200 & {
  headers: Headers;
};
export type completeEpisodeResponseError = completeEpisodeResponse400 & {
  headers: Headers;
};

export type completeEpisodeResponse =
  | completeEpisodeResponseSuccess
  | completeEpisodeResponseError;
// ...
export type toggleFavoriteEpisodeResponse200 = {
  data: ToggleFavoriteEpisode200;
  status: 200;
};

export type toggleFavoriteEpisodeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type toggleFavoriteEpisodeResponseSuccess =
  toggleFavoriteEpisodeResponse200 & {
    headers: Headers;
  };
export type toggleFavoriteEpisodeResponseError =
  toggleFavoriteEpisodeResponse400 & {
    headers: Headers;
  };

export type toggleFavoriteEpisodeResponse =
  | toggleFavoriteEpisodeResponseSuccess
  | toggleFavoriteEpisodeResponseError;
// ...
export type myReviewResponse200 = {
  data: ResultReviewResponse;
  status: 200;
};

export type myReviewResponse400 = {
  data: ResultObject;
  status: 400;
};

export type myReviewResponseSuccess = myReviewResponse200 & {
  headers: Headers;
};
export type myReviewResponseError = myReviewResponse400 & {
  headers: Headers;
};

export type myReviewResponse = myReviewResponseSuccess | myReviewResponseError;