// src/api/__generated__/story/model/postCommentRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

/**
 * 发表根评论请求
 */
export interface PostCommentRequest {
  /**
   * 评论正文，1~200 个 Unicode 字符（emoji 计 1）
   * @minLength 1
   * @maxLength 200
   */
  content: string;
}