// src/api/__generated__/story/public-comment/public-comment.ts
export type listRootCommentsResponse200 = {
  data: PageDtoCommentResponse;
  status: 200;
};

export type listRootCommentsResponse400 = {
  data: ResultObject;
  status: 400;
};

export type listRootCommentsResponseSuccess = listRootCommentsResponse200 & {
  headers: Headers;
};
export type listRootCommentsResponseError = listRootCommentsResponse400 & {
  headers: Headers;
};

export type listRootCommentsResponse =
  | listRootCommentsResponseSuccess
  | listRootCommentsResponseError;