// src/api/__generated__/story/public-drama/public-drama.ts
export type listDramaEpisodesResponse200 = {
  data: PageDtoDramaEpisodeListItemResponse;
  status: 200;
};

export type listDramaEpisodesResponse400 = {
  data: ResultObject;
  status: 400;
};

export type listDramaEpisodesResponseSuccess = listDramaEpisodesResponse200 & {
  headers: Headers;
};
export type listDramaEpisodesResponseError = listDramaEpisodesResponse400 & {
  headers: Headers;
};

export type listDramaEpisodesResponse =
  | listDramaEpisodesResponseSuccess
  | listDramaEpisodesResponseError;
// ...
export type getDramaDetailResponse200 = {
  data: DramaDetailInfoResponse;
  status: 200;
};

export type getDramaDetailResponse400 = {
  data: ResultObject;
  status: 400;
};

export type getDramaDetailResponseSuccess = getDramaDetailResponse200 & {
  headers: Headers;
};
export type getDramaDetailResponseError = getDramaDetailResponse400 & {
  headers: Headers;
};

export type getDramaDetailResponse =
  | getDramaDetailResponseSuccess
  | getDramaDetailResponseError;
// ...
export type getEpisodeDetailByEpisodeIdResponse200 = {
  data: DramaPlayResponse;
  status: 200;
};

export type getEpisodeDetailByEpisodeIdResponse400 = {
  data: ResultObject;
  status: 400;
};

export type getEpisodeDetailByEpisodeIdResponseSuccess =
  getEpisodeDetailByEpisodeIdResponse200 & {
    headers: Headers;
  };
export type getEpisodeDetailByEpisodeIdResponseError =
  getEpisodeDetailByEpisodeIdResponse400 & {
    headers: Headers;
  };

export type getEpisodeDetailByEpisodeIdResponse =
  | getEpisodeDetailByEpisodeIdResponseSuccess
  | getEpisodeDetailByEpisodeIdResponseError;