// src/api/__generated__/story/public-shortvideo/public-shortvideo.ts
export type getShortVideoDetailResponse200 = {
  data: DramaPlayResponse;
  status: 200;
};

export type getShortVideoDetailResponse400 = {
  data: ResultObject;
  status: 400;
};

export type getShortVideoDetailResponseSuccess =
  getShortVideoDetailResponse200 & {
    headers: Headers;
  };
export type getShortVideoDetailResponseError =
  getShortVideoDetailResponse400 & {
    headers: Headers;
  };

export type getShortVideoDetailResponse =
  | getShortVideoDetailResponseSuccess
  | getShortVideoDetailResponseError;