// src/api/__generated__/story/user-comment/user-comment.ts
export type postCommentResponse200 = {
  data: CommentResponse;
  status: 200;
};

export type postCommentResponse400 = {
  data: ResultObject;
  status: 400;
};

export type postCommentResponseSuccess = postCommentResponse200 & {
  headers: Headers;
};
export type postCommentResponseError = postCommentResponse400 & {
  headers: Headers;
};

export type postCommentResponse =
  | postCommentResponseSuccess
  | postCommentResponseError;
// ...
export type likeResponse200 = {
  data: CommentLikeActionResponse;
  status: 200;
};

export type likeResponse400 = {
  data: ResultObject;
  status: 400;
};

export type likeResponseSuccess = likeResponse200 & {
  headers: Headers;
};
export type likeResponseError = likeResponse400 & {
  headers: Headers;
};

export type likeResponse = likeResponseSuccess | likeResponseError;